* array.c (rb_ary_uniq_bang): do not push frozen string from hash   table.  [ruby-dev:24695]
* array.c (rb_ary_and): ditto.

* array.c (rb_ary_or): ditto.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@7190 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ regexec.c
@@ -254,7 +254,7 @@ onig_region_copy(OnigRegion* to, OnigRegion* from)
 
 /** stack **/
 #define INVALID_STACK_INDEX   -1
-typedef int  StackIndex;
+typedef long  StackIndex;
 
 typedef struct _StackType {
   unsigned int type;
